CR: Refactor Date utils
