+++ package.json
@@ -30,6 +30,7 @@
     "autoprefixer": "^10.4.16",
     "clsx": "^2.1.0",
     "crypto-js": "^4.2.0",
+    "date-fns": "^3.6.0",
     "decap-cms-app": "^3.1.10",
     "frontmatter-markdown-loader": "^3.7.0",
     "gray-matter": "^4.0.3",
